rename directory  to kebab-case
