+++ src/components/App/app.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-import style from './App.module.css'
+import style from './app.module.css'
 
-import AppHeader from '../AppHeader/AppHeader';
-import BurgerIngredients from '../BurgerIngredients/BurgerIngredients';
-import BurgerConstructor from '../BurgerConstructor/BurgerConstructor';
+import AppHeader from '../app-header/app-header';
+import BurgerIngredients from '../burger-ingredients/burger-ingredients';
+import BurgerConstructor from '../burger-constructor/burger-constructor';
 
 import {data} from '../../utils/data';
